pyproject.toml
[tool.poetry]
name = "sqlalchemy-m2m"
version = "0.1.0"
description = ""
authors = ["Naoya Yamashita <conao3@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
sqlalchemy-m2m = "sqlalchemy_m2m.main:main"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = {extras = ["postgresql"], version = "^2.0.11"}
pydantic = {extras = ["dotenv"], version = "^1.10.7"}
pyyaml = "^6.0"
sqlparse = "^0.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
